+++ test/suite/extension.test.ts
@@ -21,7 +21,7 @@ const INVALID_RULESET_PATH = path.join(__dirname, '..', '..', '..', 'rulesets', 
 const TEST_ASSETS_PATH = path.join(__dirname, '..', '..', '..', 'test', 'assets');
 const TEST_APEX_PATH = path.join(TEST_ASSETS_PATH, 'test.cls');
 
-const outputChannel = vscode.window.createOutputChannel('Apex PMD');
+const outputChannel = vscode.window.createOutputChannel('Apex PMD', {log: true});
 
 suite('Apex PMD Tests', () => {
   test('check default paths', () => {