+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Bump minimum version of VS Code API from 1.43.0 to 1.74.0 by @adangel in [#345](https://github.com/ChuckJonas/vscode-apex-pmd/pull/345)
 - Bump PMD from 7.14.0 to 7.15.0 by @adangel in [#346](https://github.com/ChuckJonas/vscode-apex-pmd/pull/346)
+- New configuration property "apexPMD.enableDebugOutput" by @adangel in [#347](https://github.com/ChuckJonas/vscode-apex-pmd/pull/347)
 ### Changed
+- Output Channel "Apex PMD" is now a LogOutputChannel and supports different log levels by @adangel in [#347](https://github.com/ChuckJonas/vscode-apex-pmd/pull/347)
 ### Deprecated
 ### Removed
 - Support for VS Code versions older than November 2022 has been removed via [#345](https://github.com/ChuckJonas/vscode-apex-pmd/pull/345)